+++ src/Currencies.php
@@ -13,10 +13,10 @@ namespace Pronamic\WordPress\Money;
 /**
  * Currencies
  *
- * @see https://github.com/moneyphp/money/blob/v3.1.3/resources/currency.php
+ * @link https://github.com/moneyphp/money/blob/v3.1.3/resources/currency.php
  *
  * @author  Remco Tolsma
- * @version 1.0.0
+ * @version 1.2.0
  * @since   1.0.0
  */
 class Currencies {
@@ -62,7 +62,7 @@ class Currencies {
 	/**
 	 * Load currencies.
 	 *
-	 * @see https://github.com/moneyphp/money/blob/v3.1.3/src/Currencies/ISOCurrencies.php#L90-L102
+	 * @link https://github.com/moneyphp/money/blob/v3.1.3/src/Currencies/ISOCurrencies.php#L90-L102
 	 * @return array
 	 *
 	 * @throws \RuntimeException Throws runtime exception if currencies could not be loaded from file.