+++ CHANGELOG.md
@@ -7,11 +7,17 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [1.2.0] - 2018-12-10
+- Added a `get_cents` method.
+- Introduced a TaxedMoney class.
+- Renamed `amount` to `value`.
+
 ## [1.1.0] - 2018-08-16
 - Added a money parser class.
 
 ## 1.0.0
 - First release.
 
-[unreleased]: https://github.com/pronamic/wp-money/compare/1.1.0...HEAD
+[unreleased]: https://github.com/pronamic/wp-money/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/pronamic/wp-money/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/pronamic/wp-money/compare/1.0.0...1.1.0